#8704 - Moves './src/dataMap.js' into './src/dataMap' directory. Adjusts paths in unit tests and mentions in comments.
