+++ src/dataMap/index.js
@@ -1,4 +1,4 @@
-import DataMap from '../dataMap';
+import DataMap from './dataMap';
 import MetaManager from './metaManager';
 import metaSchemaFactory from './metaManager/metaSchema';
 import { DynamicCellMetaMod } from './metaManager/mods/dynamicCellMeta';
